refactor: stats tracker
-Create a hook to manage dependencies common across all test types for stats tracking purposes. -Create a utility function to generate 'calculations' or strings of numbers and operators. -Add condition to ignore 'Shift' key in input accuracy tracker hook. -Add hook to manage weekly date shift logic.
